+++ src/zandronum/src/g_hexen/a_dragon.cpp
@@ -68,8 +68,8 @@ static void DragonSeek (AActor *actor, angle_t thresh, angle_t turnMax)
 	if (actor->z+actor->height < target->z ||
 		target->z+target->height < actor->z)
 	{
-		dist = P_AproxDistance(target->x-actor->x, target->y-actor->y);
-		dist = dist/actor->Speed;
+		dist = (int)(P_AproxDistance(target->x-actor->x, target->y-actor->y));
+		dist = (int)(dist/actor->Speed);
 		if (dist < 1)
 		{
 			dist = 1;
@@ -78,8 +78,8 @@ static void DragonSeek (AActor *actor, angle_t thresh, angle_t turnMax)
 	}
 	else
 	{
-		dist = P_AproxDistance (target->x-actor->x, target->y-actor->y);
-		dist = dist/actor->Speed;
+		dist = (int)(P_AproxDistance (target->x-actor->x, target->y-actor->y));
+		dist = (int)(dist/actor->Speed);
 	}
 	// [BB] If we're the server, update the thing's velocity and angle.
 	// Unfortunately there are sync issues, if we don't also update the actual position.
