+++ src/zandronum/src/g_strife/a_templar.cpp
@@ -23,7 +23,7 @@ DEFINE_ACTION_FUNCTION(AActor, A_TemplarAttack)
 
 	S_Sound (self, CHAN_WEAPON, "templar/shoot", 1, ATTN_NORM, true);	// [CW] Inform the clients.
 	A_FaceTarget (self);
-	pitch = P_AimLineAttack (self, self->angle, MISSILERANGE);
+	pitch = (int)(P_AimLineAttack (self, self->angle, MISSILERANGE));
 
 	for (int i = 0; i < 10; ++i)
 	{
